# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AgencyOS API
servers:
  - url: /api
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Auth token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
...
    User:
      type: object
      required: [id, name, email, systemRole]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        systemRole:
          type: string
        department:
          type: ["string", "null"]
        isActive:
          type: ["boolean", "null"]
    UserInput:
      type: object
      required: [name, email, systemRole]
      properties:
        name:
          type: string
        email:
          type: string
        systemRole:
          type: string
        password:
          type: ["string", "null"]
        department:
          type: ["string", "null"]
        isActive:
          type: ["boolean", "null"]
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        systemRole:
          type: string
        password:
          type: ["string", "null"]
        department:
          type: ["string", "null"]
        isActive:
          type: ["boolean", "null"]
...