# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AgencyOS API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: string;
  name: string;
  email: string;
  systemRole: string;
  /** @nullable */
  department?: string | null;
  /** @nullable */
  isActive?: boolean | null;
}