# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AgencyOS API
 * OpenAPI spec version: 0.1.0
 */

export interface UserUpdate {
  name?: string;
  email?: string;
  systemRole?: string;
  /** @nullable */
  password?: string | null;
  /** @nullable */
  department?: string | null;
  /** @nullable */
  isActive?: boolean | null;
}